1, modify map desc 2, add map desc logic
vet/app: me

+++ trunk/FTServer/Src/RegionServer/region_player_context.h
@@ -97,7 +97,7 @@ public:
 	CRITICAL_SECTION m_csContext;
 
 	// map info, need thread control
-	uint32 m_iMapId;
+	uint16 m_iMapId;
 	Map* m_pMap;
 	RegionLogicLoop* m_pLogicLoop;
 	Avatar* m_pAvatar;
